@@ -1,10 +1,10 @@
 name: listanything
 description: List anything, sometimes with a map
-version: 0.0.12+0000001
+version: 0.0.13+0000001
 publish_to: none
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
   cloud_firestore: ^4.4.3
@@ -21,9 +21,12 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_map: ^3.1.0
+  flutter_riverpod: ^2.3.6
   form_builder_validators: ^8.4.0
+  freezed_annotation: ^2.2.0
   go_router: ^6.0.9
   hooks_riverpod: ^2.3.4
+  http: ^0.13.5
   intl: ^0.17.0
   latlong2: ^0.8.1
   riverpod: ^2.2.0
@@ -35,13 +38,14 @@ dependencies:
 dev_dependencies:
   build_runner: ^2.3.0
   build_verify: ^3.1.0
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
   freezed: ^2.2.0
   go_router_builder: ^1.1.1
   json_serializable: ^6.5.4
   mocktail: ^0.3.0
-  very_good_analysis: ^2.4.0
+  very_good_analysis: ^4.0.0+1
 
 flutter:
   uses-material-design: true
