@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:listanything/app/pages/list_items/addEditListItem/add_edit_list_item.dart';
+import 'package:listanything/app/pages/list_items/upsertListItem/edit_list_item_form.dart';
 
 @immutable
 class EditListItemRoute extends GoRouteData {
@@ -13,6 +13,6 @@ class EditListItemRoute extends GoRouteData {
 
   @override
   Widget build(BuildContext context, GoRouterState state) {
-    return AddEditListItem(publicListId: publicListId, listItemId: listItemId);
+    return EditListItemForm(publicListId: publicListId, listItemId: listItemId);
   }
 }
