@@ -4,7 +4,11 @@ import 'package:listanything/app/pages/map/searchLocation/search_location_page.d
 
 @immutable
 class SearchLocationForAddPageRoute extends GoRouteData {
-  const SearchLocationForAddPageRoute({this.searchPhrase, required this.publicListId, this.listItemId});
+  const SearchLocationForAddPageRoute({
+    required this.publicListId,
+    this.listItemId,
+    this.searchPhrase,
+  });
   final String? searchPhrase;
   final String publicListId;
   final String? listItemId;
