@@ -3,7 +3,11 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:listanything/app/widgets/standardWidgets/exception_widget.dart';
 
 class AsyncValueWidget<T> extends StatelessWidget {
-  const AsyncValueWidget({super.key, required this.value, required this.data});
+  const AsyncValueWidget({
+    required this.value,
+    required this.data,
+    super.key,
+  });
   final AsyncValue<T> value;
   final Widget Function(T) data;
 
