@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 
 class Shimmer extends StatefulWidget {
   const Shimmer({
-    super.key,
     required this.linearGradient,
     this.child,
+    super.key,
   });
   static ShimmerState? of(BuildContext context) {
     return context.findAncestorStateOfType<ShimmerState>();
@@ -24,7 +24,8 @@ class ShimmerState extends State<Shimmer> with SingleTickerProviderStateMixin {
         stops: widget.linearGradient.stops,
         begin: widget.linearGradient.begin,
         end: widget.linearGradient.end,
-        transform: _SlidingGradientTransform(slidePercent: _shimmerController.value),
+        transform:
+            _SlidingGradientTransform(slidePercent: _shimmerController.value),
       );
 
   @override
@@ -42,7 +43,8 @@ class ShimmerState extends State<Shimmer> with SingleTickerProviderStateMixin {
   }
 
   Listenable get shimmerChanges => _shimmerController;
-  bool get isSized => (context.findRenderObject() as RenderBox?)?.hasSize ?? false;
+  bool get isSized =>
+      (context.findRenderObject() as RenderBox?)?.hasSize ?? false;
 
   Size get size => (context.findRenderObject()! as RenderBox).size;
 
